Some changes on the offline part
